[master] Add check on event PRINTER_STATUS_CHANGED
This event is used to check wether printer has started printing.
Once printer has started printing, a boolean is set to true, enabling
the start and/or set of the Time Trigger.
The Time Trigger is responsible of sending the pause command when no
filament has been detected for an amount of time larger than the time
threshold.